feat(core): host-owned concurrent hello app-link (KEL-30)
Lift EchoServer into keld-core and keep authenticated echo + supervised Bun
alive across the hello window via run_return, replacing sequential CLI
echo-then-window for the shipping keld dev path.
