+++ docs/agents/learnings.md
@@ -30,7 +30,8 @@ small — it is loaded by every agent session.
 - 2026-08-12 [wv/macos] Unit tests must not call `WkWebViewEngine::new` / tao `EventLoop::new` (starts AppKit). Smoke title/HTML/spec instead; `just hello` is the GUI check. (evidence: KEL-26)
 - 2026-08-12 [wv/macos] tao 0.35 `EventLoopExtMacOS` already defaults to `NSApplicationActivationPolicyRegular`; do not set it again. (evidence: tao-0.35.3 `platform/macos.rs` trait docs)
 - 2026-08-13 [ipc] postcard `from_bytes` ignores trailing bytes; a no-op echo test must assert re-encoded output ≠ input, not expect a codec error. (evidence: crates/keld-ipc/src/echo.rs `echo_roundtrip_copies_fields_not_input_bytes`)
-- 2026-08-13 [cli] tao `EventLoop::run` never returns, so `keld dev` must reap the Bun echo child before opening the hello window or the child is orphaned. (evidence: crates/keld-wv/src/wkwebview/mod.rs `run_until_closed`, crates/keld-cli/src/dev.rs `run_dev_echo` then `open_dev_window`)
+- 2026-08-21 [cli] tao `run_return` lets `keld dev` keep the supervised Bun echo child alive across the hello window and reap it after close; do not reintroduce sequential echo-then-window. (evidence: crates/keld-wv/*/run_until_closed, crates/keld-core/src/hello_session.rs, KEL-30)
+- 2026-08-13 [cli] tao `EventLoop::run` never returns (process exit); hello backends use `run_return` so Drop/reap can run after the last window closes. (evidence: crates/keld-wv/src/wkwebview/mod.rs `run_until_closed`, KEL-30)
 - 2026-08-13 [cli] rmcp 3.1.2 accepts `server/discover` only after `initialize`, with namespaced `io.modelcontextprotocol/*` metadata keys; unnamespaced keys return JSON-RPC -32602. (evidence: crates/keld-cli/tests/doctor_mcp.rs `mcp_server_discover_advertises_versions_and_identity`)
 - 2026-08-13 [ci] cargo-deny MPL-2.0 is a per-crate pin (`option-ext@0.2.0` via wry→dirs→dirs-sys), not a global allow; packed binaries must offer that crate's source. (evidence: deny.toml, docs/engineering/third-party-licenses.md, KEL-54)
 - 2026-08-13 [deps] wry crate is Apache-2.0 OR MIT; GitHub's license API reports Apache-2.0 only. tao is Apache-2.0 only. Do not treat MPL as wry's license. (evidence: crates.io wry 0.55.1 LICENSE.spdx, tao 0.35.3 Cargo.toml.orig)
@@ -41,7 +42,7 @@ small — it is loaded by every agent session.
 - 2026-08-13 [cli] Spec-named verbs that are not live (`build`/`migrate`/`gen`/`ext`) must be `KELD-CLI-045` exit 2, not a bare `unknown command` exit 1 — README agents otherwise have no tracking issue. (evidence: crates/keld-cli/src/verb.rs, KEL-29)
 - 2026-08-13 [cli] `keld create --template` must be `KELD-CLI-044` (closed flag set), not `KELD-CLI-020` invalid name `--template`; extra `keld dev` tokens must fail before tao `EventLoop::run`. (evidence: crates/keld-cli/src/flags.rs, KEL-29)
 - 2026-08-13 [process] `docs/research/` is local-only; tracking it ships 50+ notes in the PR. `git rm --cached` + gitignore, never `git add docs/research`. (evidence: PR #2, `088efc0`)
-- 2026-08-13 [cli] Unix echo sockets in shared `$TMPDIR` are world-connectable; bind inside a unique `0o700` session dir and remove the dir on Drop/shutdown. (evidence: crates/keld-cli/src/echo_link.rs `bind_unix_echo`)
+- 2026-08-13 [cli] Unix echo sockets in shared `$TMPDIR` are world-connectable; bind inside a unique `0o700` session dir and remove the dir on Drop/shutdown. (evidence: crates/keld-core/src/echo_link.rs, formerly keld-cli)
 - 2026-08-13 [ci] dtolnay/rust-toolchain@6c977a6 (2026-08-05) requires `with: toolchain:` even when rust-toolchain.toml exists; omitting it fails every job with 'toolchain' is a required input. (evidence: GitHub CI on PR #2 @ 6ed01b1)
 - 2026-08-13 [ci] Windows clippy `-D warnings` fails on unused `std::sync::mpsc` in keld-ipc `echo_link` tests; gate Unix-only test imports with `cfg(unix)`. (evidence: crates/keld-ipc/tests/echo_link.rs, PR #2 windows-latest)
 - 2026-08-13 [ci] rustdoc `-D warnings` fails on public docs linking private items (`rustdoc::private-intra-doc-links`); backtick error codes instead of `[`const`]` when the module is private. (evidence: crates/keld-cli/src/mcp/permissions.rs, PR #2 ubuntu rustdoc)
